+++ crates/cli/src/commands/filters.rs
@@ -3,14 +3,11 @@ use std::path::PathBuf;
 use anyhow::Result;
 use clap::Args;
 
-use grit_util::FileRange;
+use grit_util::{FileRange, UtilRange};
 use marzano_util::diff::FileDiff;
 use serde::Serialize;
 
-use crate::{
-    community::parse_eslint_output,
-    diff::{extract_target_diffs, extract_target_ranges},
-};
+use crate::{community::parse_eslint_output, diff::extract_target_diffs};
 
 #[derive(Args, Debug, Serialize, Default)]
 /// Shared arguments for apply and check commands.
@@ -39,7 +36,29 @@ pub(crate) fn extract_filter_ranges(
         let json_ranges = parse_eslint_output(json_content)?;
         Ok(Some(json_ranges))
     } else {
-        Ok(extract_target_ranges(&args.only_in_diff, root)?)
+        let raw_diff = extract_target_diffs(&args.only_in_diff, root)?;
+        let Some(raw_diff) = raw_diff else {
+            log::info!("No diff found, skipping range extraction");
+            return Ok(None);
+        };
+        Ok(Some(
+            raw_diff
+                .into_iter()
+                .flat_map(|diff| match diff.new_path {
+                    Some(new_path) => {
+                        let mapped = diff.ranges.into_iter().map(|range| FileRange {
+                            range: UtilRange::RangeWithoutByte(range.after),
+                            file_path: PathBuf::from(&new_path),
+                        });
+                        mapped.collect::<Vec<_>>()
+                    }
+                    None => {
+                        log::info!("Skipping diff with no new path: {:?}", diff);
+                        vec![]
+                    }
+                })
+                .collect(),
+        ))
     }
 }
